rearrange and package project for installation
Right now, I'm thinking I want to have the motor control module be
installed as a package. Multiple packages will be required for the
experiments, mainly the pan-tilt-controller, the digitizer, and the
motor control wrapper. The pan-tilt-controller and digitizer are in
separate repostiories as those are standalone packages that can be
useful outside of the drone experiments; however, the motor control
wrapper is unlikely to be useful outside of the experiments, so it
exists in this repository. There might end up being multiple packages in
this repository, so that is how I've structured it right now; this is
subject to change.